clean the websocket prefix and the helper functions in websocket.go
